requirement.txt
altair==5.0.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.42.0
gitdb==4.0.10
GitPython==3.1.32
humanfriendly==10.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.1
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
lazy_loader==0.3
llvmlite==0.40.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
onnxruntime==1.15.1
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
platformdirs==3.10.0
pooch==1.7.0
protobuf==4.24.0
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==12.0.1
pydeck==0.8.0
Pygments==2.16.1
PyMatting==1.1.8
Pympler==1.0.1
pyparsing==3.0.9
pyreadline3==3.4.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.2
rembg==2.0.50
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
scikit-image==0.21.0
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
streamlit==1.25.0
sympy==1.12
tenacity==8.2.2
tifffile==2023.7.18
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ultralytics==8.0.154
urllib3==2.0.4
validators==0.21.2
watchdog==3.0.0
zipp==3.16.2
